backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
firebase-admin
langchain
langchain-google-genai
python-dotenv
google-generativeai
